+++ services.py
@@ -1,0 +1,27 @@
+import gzip
+from gzip import GzipFile
+import io
+import json
+from six import BytesIO
+
+# Functions from getsentry/sentry-python
+
+def decompress_gzip(bytes_encoded_data):
+    try:
+        fp = BytesIO(bytes_encoded_data)
+        try:
+            f = GzipFile(fileobj=fp)
+            return f.read().decode("utf-8")
+        finally:
+            f.close()
+    except Exception as e:
+        raise e
+
+def compress_gzip(dict_body):
+    try:
+        body = io.BytesIO()
+        with gzip.GzipFile(fileobj=body, mode="w") as f:
+            f.write(json.dumps(dict_body, allow_nan=False).encode("utf-8"))
+    except Exception as e:
+        raise e
+    return body
